enh(list-manager): replace instantsearch components with algoliasearch, revert to single-file structure
