+++ package.json
@@ -42,7 +42,6 @@
     "react-dom": "^19.2.4",
     "react-hook-form": "^7.53.0",
     "react-hot-toast": "^2.5.1",
-    "react-instantsearch": "^7.32.3",
     "react-player": "^3.3.2",
     "react-truncate-list": "^1.0.3",
     "tailwindcss": "^3.4.17",
